move home to pages foldet
